Increase FastCGI buffer size.
With only 128k, the buffers would intermittenly fill up and the page output would be truncated.

The root cause for this may actually be that the temp folder isn't writable, but there're no errors in the log
to indicate that, so this fix is good enough for me for now.
See http://serverfault.com/questions/269173/why-is-the-response-truncated-at-16k-with-php-fastcgi